Ability to select files by glob when downloading a bundle
This is in preparation of manifest downloads which will limit downloading to specific files while retaining the file system layout produced by a regular bundle download.

Threre are separate sets of globs for files and metadata files because we would like to have the ability to say "Download this data file and all metadata files" and would like to do so without depending on the implementation detail that indexed files (aka metadata files) have a name that ends in `*.json`.