[CVE-2018-18074] Upgrade requests to 2.20.0 (#204)
* Upgrade requests to 2.20.0

CVE-2018-18074 More information
moderate severity
Vulnerable versions: <= 2.19.1
Patched version: 2.20.0

The Requests package through 2.19.1 before 2018-09-14 for Python sends an HTTP
Authorization header to an http URI upon receiving a same-hostname
https-to-http redirect, which makes it easier for remote attackers to discover
credentials by sniffing the network.

* Bump hca version to 4.4.9

* Pin version of flake8

* Pin moto version to 1.3.3

* Make stub param optional

For some versions of python a stub is passed into the test. In others,
it is not passed in. Since the parameter is not used, this commit makes
it optional.

* Revisions from Marcus' feedback