Add download_dir option (fixes #308)
There are a couple goals here. The first is with respect to the failing
allspark tests. The issue seemed to be because the /tmp dir was in a
separate partition from the filestore and this caused linking to fail
when the dest_path option was specified. The sensible solution would be
to place the filestore root in the same directory as dest_path. But
dest_path was poorly named. So I changed it to download_dir. The other
download funcs could use something similar I ended up solving an
auxilary goal: adding a configurable download parameter all of the
download endpoints.

The download_dir option is currently undocumented because it will be
simpler to take care of during breaking API changes in #297