Add progress bar to `hca dss upload` (closes #117)
Adds a progress bar to `hca dss upload`. This introduces a new dependency,
`tqdm` (which already exists as an indirect development requirement). The
progress bar will show by default unless one of the following conditions
is true:
* `hca` has been invoked non-interactively (i.e. stdout is not a TTY)
* The logging level is higher than INFO
* The `--no-progress` flag is passed

The progress bar displays the amount of data uploaded (instead of files
uploaded) and can guess how much time remains in the upload.